fix: merge useHealthMetrixRange.ts into correctly-spelled useHealthMetricsRange.ts
The real ~75-line useHealthMetricsRange hook implementation lived in a
misspelled file (useHealthMetrixRange.ts), with useHealthMetricsRange.ts
acting as a two-line re-export shim. This merges the implementation
directly into useHealthMetricsRange.ts and deletes the redundant
misspelled shim file. No import changes were needed since
CommitmentHealthMetrics.tsx and the existing test file already imported
from the correctly-spelled path.

closes #1227
